Project structure refactor

+++ src/cache/cache-service.js
@@ -3,27 +3,35 @@ import {
     client as cache
 } from '../database/redis'
 
+import config from './config'
 class CacheService {
 
     constructor(ttl) {
         this.ttl = ttl;
         this.pipeline = cache.pipeline;
+        this.enabled = config.enabled;
     }
 
     async get(key) {
-
-
-
-        // return await cache.get(key).then(t => {
-        //     return t;
-        // }).catch(e => {
-        //     console.error(e);
-        //     return null;
-        // });
+        if(!this.enabled){
+            console.log('Cache service is disabled.')            
+            return null;
+        }
+        return await cache.get(key).then(t => {
+            return t;
+        }).catch(e => {
+            console.error(e);
+            return null;
+        });
         return null;
     }
 
     async set(key, value) {
+        if(!this.enabled){
+            console.log('Cache service is disabled.')
+            return 0;
+        }
+
         return await cache.multi()
             .set(key, value)
             .expire(key, this.ttl)
@@ -37,10 +45,19 @@ class CacheService {
     }
 
     flush() {
+        if(!this.enabled){
+            console.log('Cache service is disabled.')
+            return 0;
+        }
         cache.flushall();
     }
 
     async remove(key) {
+        if(!this.enabled){
+            console.log('Cache service is disabled.')
+            return 0;
+        }
+
         await cache.del(key).then(d => {
             return 1;
         }).catch(e => {
